Fix problem with extra characters in subtitle
The .caption files consist of subtitle text and timestamps with some
binary data interposed. When converted to text, this binary data
produces random, mostly unprintable characters. Many can be filtered
out, but printable characters are sometimes overlooked and end up in the
subtitle. This problem can be overcome by taking a different approach.
Study revealed that the subtitle text and timestamp data appears at
predictable positions in the .caption file. This commit replaces the
filter method with one that makes use of this principle, extracting only
text and timestamps and leaving out any binary data.